ripristinati tutti gli id a LONG

+++ AccelerAudio/src/team/bugbusters/acceleraudio/PlayRecord.java
@@ -16,7 +16,7 @@ import android.os.SystemClock;
 public class PlayRecord extends IntentService {
 	private DbAdapter dbHelper;
 	private Cursor cr;
-	private int id_to_process;
+	private long id_to_process;
 	private String asseX,asseY,asseZ;
 	private boolean checkX,checkY,checkZ;
 	private int sovrac;
@@ -54,7 +54,7 @@ public class PlayRecord extends IntentService {
         }
         
         dbHelper = new DbAdapter(this);
-        id_to_process=intent.getIntExtra("ID", -1);
+        id_to_process=intent.getLongExtra("ID", -1);
         
         /*-- Opening DB --*/
         dbHelper.open();
